[PA] Modified FreeSlotBitmapCellType to uint64_t
FreeSlotBitmapCellType was uintptr_t, but this might be problematic
because it's architecture dependent. This CL changed it to uint64_t.

Bug: 1354062
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3869968
Commit-Queue: Honoka Anada <honokaanada@google.com>
Cr-Commit-Position: refs/heads/main@{#1042497}

+++ base/allocator/partition_allocator/freeslot_bitmap_constants.h
@@ -15,7 +15,7 @@
 
 namespace partition_alloc::internal {
 
-using FreeSlotBitmapCellType = uintptr_t;
+using FreeSlotBitmapCellType = uint64_t;
 constexpr size_t kFreeSlotBitmapBitsPerCell =
     sizeof(FreeSlotBitmapCellType) * CHAR_BIT;
 constexpr size_t kFreeSlotBitmapOffsetMask = kFreeSlotBitmapBitsPerCell - 1;
